# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  controllers.RegisterRequest:
    properties:
      display_name:
        maxLength: 100
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  middleware.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  middleware.ResponseEnvelope:
    properties:
      api_version:
        type: string
      data: {}
      error:
        $ref: '#/definitions/middleware.ErrorInfo'
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@travellink.local
    name: API Support
    url: http://www.swagger.io/support
  description: Travel Linked List API - Browse, create, and share travel plans
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TravelLink API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password to receive JWT access token
      parameters:
      - description: Login request with email and password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully authenticated
          schema:
            $ref: '#/definitions/middleware.ResponseEnvelope'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/middleware.ResponseEnvelope'
        "401":
          description: Invalid credentials or inactive user
          schema:
            $ref: '#/definitions/middleware.ResponseEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ResponseEnvelope'
      summary: Authenticate user and get access token
      tags:
      - auth
  /auth/logout:
    post:
      description: Logout authenticated user. In a stateless JWT system, the client
        should discard the token.
      responses:
        "204":
          description: User successfully logged out
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/middleware.ResponseEnvelope'
      security:
      - Bearer: []
      summary: Logout user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, username, password, and display
        name
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User successfully registered
          schema:
            $ref: '#/definitions/middleware.ResponseEnvelope'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/middleware.ResponseEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ResponseEnvelope'
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  Bearer:
    description: 'Enter JWT token in the format: Bearer eyJhbGci...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"